refact: refactoring the upload into a service
