+++ app/src/main/java/com/ensak/connect/service/ActivityResultCallback.java
@@ -1,0 +1,6 @@
+package com.ensak.connect.service;
+
+public interface ActivityResultCallback<T> {
+    void onSuccess(T data);
+    void onError(Throwable throwable);
+}
